+++ changelog.md
@@ -1,5 +1,11 @@
 ## OCHRE Changelog
 
+### Changes from PRs
+
+- Fixed bug with accounting for HVAC delivered heat for standalone HVAC runs 
+- Fixed bug with ASHP backup heater units
+- Added OCHREException class to handle errors
+
 ### OCHRE v0.8.4-beta
 
 - Fixed bug with air infiltration inputs (works with ResStock 3.0 and 3.1, and OS-HPXML 1.6.0)